add saved Trip view

+++ src/client/ts/savedPlaces.ts
@@ -1,78 +1,24 @@
 import { html, render } from 'lit-html'
 import * as _ from './helperFunctions'
 import { getSavedTrips, Trip, deleteTripById } from './apiHandler'
-import { renderSavedTripList } from './tripData'
+import { renderSavedTripList, renderSavedTrip } from './tripData'
 import { renderBestPlaces } from './bestPlaces'
 
 import '../styles/saved-places.scss'
 
-// function renderPicture(pictures: PictureData[], element: HTMLElement) {
-//   element.style.backgroundImage = 'url(' + pictures[0].previewURL + ')'
-// }
-
-// const locationsTemp = ({ location, startDate, endDate }: Location) => html`
-//   <div class="location-container" key=${location}>
-//     <div class="loction-name">${location}</div>
-//     <div class="location-date">
-//       From: ${converStringToLocaledDate(startDate)}
-//     </div>
-//     <div class="location-date">To: ${converStringToLocaledDate(endDate)}</div>
-//   </div>
-// `
-// const packingItemTemp = (packingItem: string) => html`<div>${packingItem}</div>`
-
-// const savedTripTemp = (item: string, trip: Trip) => html`
-//   <div class="trip-container" id=${item}>
-//     <h3>${item}</h3>
-//     <div class="action-items">
-//       <button id="open-${item}" @click=${() => openTrip(item, trip)}>
-//         Open Trip
-//       </button>
-//       <button id="delete-${item}" @click=${() => deleteTrip(item)}>
-//         DELETE
-//       </button>
-//     </div>
-
-//     <div class="locations-container ">
-//       ${trip.locations.length
-//         ? trip.locations.map(locationsTemp)
-//         : html`<div class="empty">No location added to this Trip yet!</div>`}
-//     </div>
-//     <div class="packing-list">
-//       <h4>Packing List:</h4>
-//       ${trip.packingList.length
-//         ? trip.packingList.map(packingItemTemp)
-//         : html`<div class="empty">No Items added to your Packlist yet!</div>`}
-//     </div>
-//   </div>
-//`
-// const renderLocationPicture = (location: string, tripName: string) => {
-//   const pictureHolder = document.querySelectorAll(
-//     `.location-container[key="${location}"]`
-//   )
-//   getPictures({ name: location, countryName: tripName }).then(picData =>
-//     pictureHolder.forEach(it => renderPicture(picData, it as HTMLElement))
-//   )
-// }
-
-const renderMain = (trip: Trip) => {
-  const data = html`<div class="trip-editor">
-    <h1>${trip.location}</h1>
-    <div class="current-saved-trips" id="currentSavedTrips"></div>
-  </div>`
-  render(data, _.main())
-}
-
 const openTrip = (item: string, trip: Trip) => {
   setActiveTripElement(item)
-  renderMain(trip)
+  renderSavedTrip(trip, _.main())
 }
 
 const deleteTrip = async (id: number) => {
-  const status = await deleteTripById(id).then(it => it.status)
-  if (status !== 202) return
-  if (_.getHashSave()['tripId'] === String(id)) history.pushState(null, '', '#')
-  await init()
+  if (confirm('Are you shore to delete this Trip ?')) {
+    const status = await deleteTripById(id).then(it => it.status)
+    if (status !== 202) return
+    if (_.getHashSave()['tripId'] === String(id))
+      history.pushState(null, '', '#')
+    await init()
+  }
 }
 
 const setEventListener = (trips: Trip[]) => {
@@ -105,7 +51,6 @@ export async function init() {
   const savedTrips = (await getSavedTrips()).sort((a, b) =>
     a.startDate > b.startDate ? 1 : -1
   )
-
   if (savedTrips.length) {
     const hashTripName = _.getHashSave()['tripId']
     const presetTripName = savedTrips.some(it => String(it.id) === hashTripName)
@@ -120,7 +65,8 @@ export async function init() {
     await renderSavedTripList(savedTrips, savedContainer)
     setActiveTripElement(presetTripName)
     setEventListener(savedTrips)
-    renderMain(savedTrips.find(it => String(it.id) === hashTripName)!)
+    const trip = savedTrips.find(it => String(it.id) === presetTripName)!
+    renderSavedTrip(trip, _.main())
   } else {
     renderBestPlaces()
     render(html`<h1 class="empty">You have not saved a Trip yet</h1>`, _.main())
